requirements.txt
azure-functions
python-jobspy>=1.1.79
pandas>=2.2.0
